added medical and medicine view
