+++ backend/pharmacy/api/views.py
@@ -37,119 +37,7 @@ from django.contrib.auth.models import User
 from rest_framework import filters
 
 
-# A method decorator to cache the view for 60 seconds
-@method_decorator(cache_page(60), name='get')
-# A method decorator to vary on the cookie
-@method_decorator(vary_on_cookie, name='get')
-# A method decorator to vary on the headers
-@method_decorator(vary_on_headers, name='get')
-# Allow only authenticated users to access this view
-# @permission_classes([IsAuthenticated])
-class MedicalView(APIView):
-    def getObject(self, pk):
-        try:
-            return Medical.objects.filter(pk=pk)
-        except Medical.DoesNotExist:
-            return Http404
 
-    def get(self, request, pk):
-        medical = self.getObject(pk)
-        serializer = MedicalSerializer(medical, many=True)
-        return Response(serializer.data)
-
-    def put(self, request, pk):
-        medical = self.getObject(pk)
-        # Ensure that the user is the owner of the medical
-        if medical[0].user.id != request.user.id:
-            return Response("HTTP 403 Forbidden", status=403)
-        serializer = MedicalSerializer(medical, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-
-    def delete(self, request, pk):
-        medical = self.getObject(pk)
-        # Ensure that the user is the owner of the medical
-        if medical[0].user.id != request.user.id:
-            return Response("HTTP 403 Forbidden", status=403)
-        medical.delete()
-        return Response("Deleted", status=200)
-
-
-# @permission_classes([IsAuthenticated])
-class MedicalViewList(APIView):
-    def get(self, request, format=None):
-        medical = Medical.objects.all()
-        serializer = MedicalSerializer(medical, many=True)
-        return Response(serializer.data)
-
-    def post(self, request):
-        serializer = MedicalSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-# @permission_classes([IsAuthenticated])
-class MedicalSearch(generics.ListCreateAPIView):
-    search_fields = ['name', 'address', 'phone']
-    filter_backends = (filters.SearchFilter,)
-    queryset = Medical.objects.all()
-    serializer_class = MedicalSerializer
-
-# @permission_classes([IsAuthenticated])
-class MedicineView(APIView):
-    def getObject(self, pk):
-        try:
-            return Medicine.objects.filter(pk=pk)
-        except Medicine.DoesNotExist:
-            return Http404
-
-    def get(self, request, pk):
-        medicine = self.getObject(pk)
-        serializer = MedicineSerializer(medicine, many=True)
-        return Response(serializer.data)
-
-    def put(self, request, pk):
-        medicine = self.getObject(pk)
-        # Ensure that the user is the owner of the medicine
-        if medicine[0].user.id != request.user.id:
-            return Response("HTTP 403 Forbidden", status=403)
-        serializer = MedicineSerializer(medicine, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-
-    def delete(self, request, pk):
-        medicine = self.getObject(pk)
-        # Ensure that the user is the owner of the medicine
-        if medicine[0].user.id != request.user.id:
-            return Response("HTTP 403 Forbidden", status=403)
-        medicine.delete()
-        return Response("Deleted", status=200)
-
-# @permission_classes([IsAuthenticated])
-class MedicineViewList(APIView):
-    def get(self, request, format=None):
-        medicine = Medicine.objects.all()
-        serializer = MedicineSerializer(medicine, many=True)
-        return Response(serializer.data)
-
-    def post(self, request):
-        serializer = MedicineSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-# @permission_classes([IsAuthenticated])
-class MedicineSearch(generics.ListCreateAPIView):
-    search_fields = ['name', 'description']
-    filter_backends = (filters.SearchFilter,)
-    queryset= Medicine.objects.all()
-    serializer_class = MedicineSerializer
