+++ backend/pharmacy/api/views/medical/search.py
@@ -8,30 +8,24 @@
 #
 # All rights reserved.
 
-from django.http import Http404
-from rest_framework.response import Response
 from rest_framework.decorators import permission_classes
 from rest_framework.permissions import IsAuthenticated
 
-from pharmacy.models import Medical, Medicine
-from ...serializers import MedicalSerializer, MedicineSerializer, RegisterSerializer
-
-# For customizing user claims
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from rest_framework_simplejwt.views import TokenObtainPairView
-
-# Imports for caching
-from rest_framework.views import APIView
-
-from django.utils.decorators import method_decorator
-from django.views.decorators.cache import cache_page
-from django.views.decorators.vary import vary_on_cookie, vary_on_headers
-
+from pharmacy.models import Medical
+from ...serializers import MedicalSerializer
 
 # Imports for registering a new user
 from rest_framework import generics
-from rest_framework.permissions import AllowAny
-from django.contrib.auth.models import User
 
 # Imports used in search functionality
 from rest_framework import filters
+
+
+
+@permission_classes([IsAuthenticated])
+class MedicalSearch(generics.ListCreateAPIView):
+    search_fields = ['name', 'address', 'phone']
+    filter_backends = (filters.SearchFilter,)
+    queryset = Medical.objects.all()
+    serializer_class = MedicalSerializer
+
