+++ backend/pharmacy/api/views/medical/views.py
@@ -13,12 +13,9 @@ from rest_framework.response import Response
 from rest_framework.decorators import permission_classes
 from rest_framework.permissions import IsAuthenticated
 
-from pharmacy.models import Medical, Medicine
-from ...serializers import MedicalSerializer, MedicineSerializer, RegisterSerializer
+from pharmacy.models import Medical
+from ...serializers import MedicalSerializer
 
-# For customizing user claims
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from rest_framework_simplejwt.views import TokenObtainPairView
 
 # Imports for caching
 from rest_framework.views import APIView
@@ -28,10 +25,41 @@ from django.views.decorators.cache import cache_page
 from django.views.decorators.vary import vary_on_cookie, vary_on_headers
 
 
-# Imports for registering a new user
-from rest_framework import generics
-from rest_framework.permissions import AllowAny
-from django.contrib.auth.models import User
+# A method decorator to cache the view for 60 seconds
+@method_decorator(cache_page(60), name='get')
+# A method decorator to vary on the cookie
+@method_decorator(vary_on_cookie, name='get')
+# A method decorator to vary on the headers
+@method_decorator(vary_on_headers, name='get')
+# Allow only authenticated users to access this view
+@permission_classes([IsAuthenticated])
+class MedicalView(APIView):
+    def getObject(self, pk):
+        try:
+            return Medical.objects.filter(pk=pk)
+        except Medical.DoesNotExist:
+            return Http404
 
-# Imports used in search functionality
-from rest_framework import filters
+    def get(self, request, pk):
+        medical = self.getObject(pk)
+        serializer = MedicalSerializer(medical, many=True)
+        return Response(serializer.data)
+
+    def put(self, request, pk):
+        medical = self.getObject(pk)
+        # Ensure that the user is the owner of the medical
+        if medical[0].user.id != request.user.id:
+            return Response("HTTP 403 Forbidden", status=403)
+        serializer = MedicalSerializer(medical, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
+
+    def delete(self, request, pk):
+        medical = self.getObject(pk)
+        # Ensure that the user is the owner of the medical
+        if medical[0].user.id != request.user.id:
+            return Response("HTTP 403 Forbidden", status=403)
+        medical.delete()
+        return Response("Deleted", status=200)
